requirements.txt
pandas
setuptools
setuptools_scm
importlib-resources
importlib_metadata
numpy
cantools>=32.20.1
ipython
vin_parser
scipy>=1.6.1
python-dotenv